+++ RELEASE_NOTES.md
@@ -31,6 +31,8 @@
 - improved: functional tests invoke `discover` and `traverse`
 - fixed: command-line argument validation for `list`
 - added: `--list-commits-with-hashes` flag to `status`
+- improved: stability of loading branch data from git
+- fixed: predictability of handling branch remote tracking data
 
 ## New in git-machete 2.12.10